Handle cleaning of port when log_file has not been initialised.
log_file gets initialised as part of the DEPEND stage, if that stage has not
been started then log_file contains the default value, of None.

+++ libpb/port/port.py
@@ -162,9 +162,8 @@ class Port(object):
         self.working = False
         if job and not job.status:
             self.failed = True
-        if not self.failed:
-            if os.path.isfile(self.log_file):
-                os.unlink(self.log_file)
+        if not self.failed and self.log_file and os.path.isfile(self.log_file):
+            os.unlink(self.log_file)
 
     def reset(self):
         """Reset the ports state, and stage."""
